Prevent timeout
If the player is idle for more than a minute, their socket times out
but they still see themselves as "in game" on their screen.

To prevent this, we get them to rejoin the game.

They will still time out after a certain point (30 reconnects = 30 mins)
but this should be plenty of time to get the game up and running.